+++ src/main/java/frc/robot/commands/vision/LimeAlign.java
@@ -2,7 +2,7 @@
 // Open Source Software; you can modify and/or share it under the terms of
 // the WPILib BSD license file in the root directory of this project.
 
-package frc.robot.commands.visionCommands;
+package frc.robot.commands.vision;
 
 import edu.wpi.first.math.controller.PIDController;
 import edu.wpi.first.wpilibj2.command.PIDCommand;
@@ -11,9 +11,9 @@ import frc.robot.RobotContainer;
 // NOTE:  Consider using this command inline, rather than writing a subclass.  For more
 // information, see:
 // https://docs.wpilib.org/en/stable/docs/software/commandbased/convenience-features.html
-public class limeAlign extends PIDCommand {
+public class LimeAlign extends PIDCommand {
   /** Creates a new limeAlign. */
-  public limeAlign() {
+  public LimeAlign() {
     super(
         // The controller that the command will use
         new PIDController(0, 0, 0),
